# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stellar Viaduct API",
    "description": "Real-time Stellar bridge monitoring API. Provides bridge statuses, asset health, liquidity, price feeds, alert management, analytics, and administrative controls.\n\n## Authentication\nProtected endpoints require an API key in the `x-api-key` header.\n\n## Rate Limiting\nAll endpoints are rate-limited per IP using a sliding-window algorithm backed by Redis. Exceeding the limit returns **429 Too Many Requests** with a `Retry-After` header.\n\n## Versioning\nAll REST endpoints are prefixed with `/api/v1/`. Breaking changes increment the version segment.\n\n## Error Format\n```json\n{ \"error\": \"Short label\", \"message\": \"Human-readable description\" }\n```",
    "version": "1.0.0",
    "contact": {
      "name": "Stellar Viaduct Team",
      "url": "https://github.com/Stellar_Viaduct/Stellar_Viaduct"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" },
    { "url": "https://api.stellarviaduct.io", "description": "Production" }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for protected endpoints"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Not Found" },
          "message": { "type": "string", "example": "The requested resource was not found" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/assets/{symbol}": {
      "get": {
        "tags": ["Assets"],
        "summary": "Get asset details",
        "parameters": [{ "name": "symbol", "in": "path", "required": true, "schema": { "type": "string" }, "example": "USDC" }],
        "responses": {
          "200": { "description": "Asset details" },
          "404": { "description": "Asset not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/v1/assets/{symbol}/liquidity": {
      "get": {
        "tags": ["Assets"],
        "summary": "Get aggregated liquidity data",
        "parameters": [{ "name": "symbol", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Liquidity data" } }
      }
    },
    "/api/v1/assets/{symbol}/price": {
      "get": {
        "tags": ["Assets"],
        "summary": "Get aggregated price from all sources",
        "parameters": [{ "name": "symbol", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Aggregated price" } }
      }
    },
...
  }
}